+++ v2/models/extender_controller_extender_profile.go
@@ -4,11 +4,11 @@ const ExtenderControllerExtenderProfilePath = "extender-controller/extender-prof
 
 type ExtenderControllerExtenderProfile struct {
 	Allowaccess         *string                                          `json:"allowaccess,omitempty"`
-	BandwidthLimit      *int64                                           `json:"bandwidth-limit,omitempty"`
+	BandwidthLimit      *float64                                         `json:"bandwidth-limit,omitempty"`
 	Cellular            *[]ExtenderControllerExtenderProfileCellular     `json:"cellular,omitempty"`
 	EnforceBandwidth    *string                                          `json:"enforce-bandwidth,omitempty"`
 	Extension           *string                                          `json:"extension,omitempty"`
-	Fosid               *int64                                           `json:"fosid,omitempty"`
+	Fosid               *float64                                         `json:"fosid,omitempty"`
 	LanExtension        *[]ExtenderControllerExtenderProfileLanExtension `json:"lan-extension,omitempty"`
 	LoginPassword       *string                                          `json:"login-password,omitempty"`
 	LoginPasswordChange *string                                          `json:"login-password-change,omitempty"`
@@ -25,9 +25,9 @@ type ExtenderControllerExtenderProfileCellular struct {
 }
 
 type ExtenderControllerExtenderProfileCellularControllerReport struct {
-	Interval        *int64  `json:"interval,omitempty"`
-	SignalThreshold *int64  `json:"signal-threshold,omitempty"`
-	Status          *string `json:"status,omitempty"`
+	Interval        *float64 `json:"interval,omitempty"`
+	SignalThreshold *float64 `json:"signal-threshold,omitempty"`
+	Status          *string  `json:"status,omitempty"`
 }
 
 type ExtenderControllerExtenderProfileCellularDataplan struct {
@@ -36,7 +36,7 @@ type ExtenderControllerExtenderProfileCellularDataplan struct {
 
 type ExtenderControllerExtenderProfileCellularModem1 struct {
 	AutoSwitch       *[]ExtenderControllerExtenderProfileCellularModem1AutoSwitch `json:"auto-switch,omitempty"`
-	ConnStatus       *int64                                                       `json:"conn-status,omitempty"`
+	ConnStatus       *float64                                                     `json:"conn-status,omitempty"`
 	DefaultSim       *string                                                      `json:"default-sim,omitempty"`
 	Gps              *string                                                      `json:"gps,omitempty"`
 	PreferredCarrier *string                                                      `json:"preferred-carrier,omitempty"`
@@ -49,19 +49,19 @@ type ExtenderControllerExtenderProfileCellularModem1 struct {
 }
 
 type ExtenderControllerExtenderProfileCellularModem1AutoSwitch struct {
-	Dataplan            *string `json:"dataplan,omitempty"`
-	Disconnect          *string `json:"disconnect,omitempty"`
-	DisconnectPeriod    *int64  `json:"disconnect-period,omitempty"`
-	DisconnectThreshold *int64  `json:"disconnect-threshold,omitempty"`
-	Signal              *string `json:"signal,omitempty"`
-	SwitchBack          *string `json:"switch-back,omitempty"`
-	SwitchBackTime      *string `json:"switch-back-time,omitempty"`
-	SwitchBackTimer     *int64  `json:"switch-back-timer,omitempty"`
+	Dataplan            *string  `json:"dataplan,omitempty"`
+	Disconnect          *string  `json:"disconnect,omitempty"`
+	DisconnectPeriod    *float64 `json:"disconnect-period,omitempty"`
+	DisconnectThreshold *float64 `json:"disconnect-threshold,omitempty"`
+	Signal              *string  `json:"signal,omitempty"`
+	SwitchBack          *string  `json:"switch-back,omitempty"`
+	SwitchBackTime      *string  `json:"switch-back-time,omitempty"`
+	SwitchBackTimer     *float64 `json:"switch-back-timer,omitempty"`
 }
 
 type ExtenderControllerExtenderProfileCellularModem2 struct {
 	AutoSwitch       *[]ExtenderControllerExtenderProfileCellularModem2AutoSwitch `json:"auto-switch,omitempty"`
-	ConnStatus       *int64                                                       `json:"conn-status,omitempty"`
+	ConnStatus       *float64                                                     `json:"conn-status,omitempty"`
 	DefaultSim       *string                                                      `json:"default-sim,omitempty"`
 	Gps              *string                                                      `json:"gps,omitempty"`
 	PreferredCarrier *string                                                      `json:"preferred-carrier,omitempty"`
@@ -74,14 +74,14 @@ type ExtenderControllerExtenderProfileCellularModem2 struct {
 }
 
 type ExtenderControllerExtenderProfileCellularModem2AutoSwitch struct {
-	Dataplan            *string `json:"dataplan,omitempty"`
-	Disconnect          *string `json:"disconnect,omitempty"`
-	DisconnectPeriod    *int64  `json:"disconnect-period,omitempty"`
-	DisconnectThreshold *int64  `json:"disconnect-threshold,omitempty"`
-	Signal              *string `json:"signal,omitempty"`
-	SwitchBack          *string `json:"switch-back,omitempty"`
-	SwitchBackTime      *string `json:"switch-back-time,omitempty"`
-	SwitchBackTimer     *int64  `json:"switch-back-timer,omitempty"`
+	Dataplan            *string  `json:"dataplan,omitempty"`
+	Disconnect          *string  `json:"disconnect,omitempty"`
+	DisconnectPeriod    *float64 `json:"disconnect-period,omitempty"`
+	DisconnectThreshold *float64 `json:"disconnect-threshold,omitempty"`
+	Signal              *string  `json:"signal,omitempty"`
+	SwitchBack          *string  `json:"switch-back,omitempty"`
+	SwitchBackTime      *string  `json:"switch-back-time,omitempty"`
+	SwitchBackTimer     *float64 `json:"switch-back-timer,omitempty"`
 }
 
 type ExtenderControllerExtenderProfileCellularSmsNotification struct {
@@ -116,8 +116,8 @@ type ExtenderControllerExtenderProfileLanExtension struct {
 }
 
 type ExtenderControllerExtenderProfileLanExtensionBackhaul struct {
-	Name   *string `json:"name,omitempty"`
-	Port   *string `json:"port,omitempty"`
-	Role   *string `json:"role,omitempty"`
-	Weight *int64  `json:"weight,omitempty"`
+	Name   *string  `json:"name,omitempty"`
+	Port   *string  `json:"port,omitempty"`
+	Role   *string  `json:"role,omitempty"`
+	Weight *float64 `json:"weight,omitempty"`
 }
