+++ v2/models/firewall.wildcard_fqdn_custom.go
@@ -3,10 +3,10 @@ package models
 const FirewallwildcardFqdnCustomPath = "firewall.wildcard-fqdn/custom/"
 
 type FirewallwildcardFqdnCustom struct {
-	Color        *int64  `json:"color,omitempty"`
-	Comment      *string `json:"comment,omitempty"`
-	Name         *string `json:"name,omitempty"`
-	Uuid         *string `json:"uuid,omitempty"`
-	Visibility   *string `json:"visibility,omitempty"`
-	WildcardFqdn *string `json:"wildcard-fqdn,omitempty"`
+	Color        *float64 `json:"color,omitempty"`
+	Comment      *string  `json:"comment,omitempty"`
+	Name         *string  `json:"name,omitempty"`
+	Uuid         *string  `json:"uuid,omitempty"`
+	Visibility   *string  `json:"visibility,omitempty"`
+	WildcardFqdn *string  `json:"wildcard-fqdn,omitempty"`
 }
